fix CI: useless cast, a semicolon in a test name, stale configuration reference
The last merges were verified with the release preset and ctest only:
FASTMM_WERROR is off there, and the lint and docs checks did not run.

+++ bench/bench_order_tcp.cpp
@@ -87,7 +87,7 @@ const Setup& setup() {
         tid = -1;
         return;
       }
-      tid = static_cast<int>(::gettid());
+      tid = ::gettid();
       while (!configured.load()) std::this_thread::yield();
       r.kernel_listen = listen_on(7100);
       r.user_listen = listen_on(7101);
